+++ src/headless/core.js
@@ -243,6 +243,11 @@ export const api = _converse.api = {
                 // Recreate all the promises
                 Object.keys(_converse.promises).forEach(replacePromise);
                 delete _converse.jid
+
+                // Remove the session JID, otherwise the user would just be logged
+                // in again upon reload. See #2759
+                localStorage.removeItem('conversejs-session-jid');
+
                 /**
                  * Triggered once the user has logged out.
                  * @event _converse#logout